+++ core/fxge/android/cfx_androidfontinfo.h
@@ -31,7 +31,7 @@ class CFX_AndroidFontInfo final : public SystemFontInfoIface {
                 int pitch_family,
                 const ByteString& face) override;
   void* GetFont(const ByteString& face) override;
-  uint32_t GetFontData(void* hFont,
+  size_t GetFontData(void* hFont,
                        uint32_t table,
                        pdfium::span<uint8_t> buffer) override;
   bool GetFaceName(void* hFont, ByteString* name) override;
